Theme enhancements to docs and new theme
Added:
- Thanksgiving2 theme. Nice dark theme as a second
option to the light one.
- Theme generation guide, to specify the details
of how to generate themes and the nuances.
- Prompt pack - docs with prompts on how to
generate theme content. Were used to generate the
new thanksgiving2 theme (and many other iterations
for testing to nail down the prompts and process).
- Word lists to go along with Thanksgiving and
Thanksgiving2 themes. (They are not tied to the
themes, but wanted to give some extra selection.)